Move error types to an `errors.rs` module.
Separates the error types so that they no longer live in the contracts
module and live in their own separate module.
