+++ substrate/frame/revive/wire/src/types/mod.rs
@@ -29,6 +29,7 @@ mod block;
 mod bytes;
 mod code;
 mod contract;
+mod errors;
 mod receipt;
 mod state_overrides;
 mod storage;
@@ -43,6 +44,7 @@ pub use block::*;
 pub use bytes::*;
 pub use code::*;
 pub use contract::*;
+pub use errors::*;
 pub use receipt::*;
 pub use state_overrides::*;
 pub use storage::*;
